# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/surveys": {
      "get": {
        "operationId": "SurveysController_findAll_v1",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search query - searches title, client name, property address, and job ref",
            "schema": {
              "maxLength": 200,
              "example": "Main Street",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (1-100)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by survey status",
            "schema": {
              "enum": [
                "DRAFT",
                "IN_PROGRESS",
                "PAUSED",
                "COMPLETED",
                "PENDING_REVIEW",
                "APPROVED",
                "REJECTED"
              ],
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by survey type",
            "schema": {
              "enum": [
                "LEVEL_2",
                "LEVEL_3",
                "VALUATION",
                "SNAGGING",
                "REINSPECTION",
                "OTHER"
              ],
              "type": "string"
            }
          },
          {
            "name": "clientName",
            "required": false,
            "in": "query",
            "description": "Filter by client name (partial match)",
            "schema": {
              "maxLength": 255,
              "example": "Smith",
              "type": "string"
            }
          },
          {
            "name": "createdFrom",
            "required": false,
            "in": "query",
            "description": "Filter surveys created on or after this date (ISO 8601)",
            "schema": {
              "example": "2024-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "createdTo",
            "required": false,
            "in": "query",
            "description": "Filter surveys created on or before this date (ISO 8601)",
            "schema": {
              "example": "2024-12-31T23:59:59.999Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of surveys matching search criteria",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - insufficient role"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Search and list surveys with pagination and filtering",
        "tags": [
          "Surveys"
        ]
      },
      "post": {
        "operationId": "SurveysController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSurveyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Survey created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - insufficient role"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new survey with sections and answers",
        "tags": [
          "Surveys"
        ]
      }
    },
    "/api/v1/surveys/{id}": {
      "get": {
        "operationId": "SurveysController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Survey UUID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Survey retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not owner or admin"
          },
          "404": {
            "description": "Survey not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get a survey by ID with all sections and answers",
        "tags": [
          "Surveys"
        ]
      },
      "put": {
        "operationId": "SurveysController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Survey UUID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSurveyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Survey updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not owner or admin"
          },
          "404": {
            "description": "Survey not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a survey with sections and answers",
        "tags": [
          "Surveys"
        ]
      },
      "delete": {
        "operationId": "SurveysController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Survey UUID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Survey soft-deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteSurveyResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not owner or admin"
          },
          "404": {
            "description": "Survey not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Soft delete a survey (marks as deleted)",
        "tags": [
          "Surveys"
        ]
      }
    },
...
  },
  "info": {
    "title": "SurveyScriber API",
    "description": "SurveyScriber Backend API - Professional Field Survey & Inspection Platform\n\n## Authentication\nThis API uses JWT Bearer tokens. Obtain tokens via `/auth/login` and include in the Authorization header.\n\n## Offline-First Architecture\nThe API supports granular CRUD operations for surveys, sections, and answers to enable offline-first Flutter clients to sync individual changes without full survey uploads.",
    "version": "1.0.0",
    "contact": {
      "name": "SurveyScriber Team",
      "url": "https://surveyscriber.com",
      "email": "support@surveyscriber.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token from /auth/login response",
        "in": "header"
      }
    },
    "schemas": {
...
      "SurveyListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "title": {
            "type": "string",
            "example": "Level 2 Home Survey - 123 Main Street"
          },
          "propertyAddress": {
            "type": "string",
            "example": "123 Main Street, London, SW1A 1AA"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "IN_PROGRESS",
              "PAUSED",
              "COMPLETED",
              "PENDING_REVIEW",
              "APPROVED",
              "REJECTED"
            ],
            "example": "DRAFT"
          },
          "type": {
            "type": "string",
            "enum": [
              "LEVEL_2",
              "LEVEL_3",
              "VALUATION",
              "SNAGGING",
              "REINSPECTION",
              "OTHER"
            ],
            "example": "LEVEL_2"
          },
          "jobRef": {
            "type": "string",
            "example": "JOB-2024-001234"
          },
          "clientName": {
            "type": "string",
            "example": "John Smith"
          },
          "userId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "title",
          "propertyAddress",
          "status",
          "userId",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1,
            "description": "Current page number"
          },
          "limit": {
            "type": "number",
            "example": 20,
            "description": "Items per page"
          },
          "total": {
            "type": "number",
            "example": 100,
            "description": "Total number of items"
          },
          "totalPages": {
            "type": "number",
            "example": 5,
            "description": "Total number of pages"
          },
          "hasNext": {
            "type": "boolean",
            "example": true,
            "description": "Whether there is a next page"
          },
          "hasPrev": {
            "type": "boolean",
            "example": false,
            "description": "Whether there is a previous page"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages",
          "hasNext",
          "hasPrev"
        ]
      },
      "SurveyListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SurveyListItemDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "CreateSectionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Roof Inspection",
            "description": "Title of the section"
          },
          "order": {
            "type": "number",
            "example": 1,
            "description": "Order of the section within the survey",
            "default": 0
          }
        },
        "required": [
          "title"
        ]
      },
      "CreateSurveyDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Level 2 Home Survey - 123 Main Street",
            "description": "Title of the survey"
          },
          "propertyAddress": {
            "type": "string",
            "example": "123 Main Street, London, SW1A 1AA",
            "description": "Property address being surveyed"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "IN_PROGRESS",
              "PAUSED",
              "COMPLETED",
              "PENDING_REVIEW",
              "APPROVED",
              "REJECTED"
            ],
            "example": "DRAFT",
            "description": "Survey status",
            "default": "DRAFT"
          },
          "type": {
            "type": "string",
            "enum": [
              "LEVEL_2",
              "LEVEL_3",
              "VALUATION",
              "SNAGGING",
              "REINSPECTION",
              "OTHER"
            ],
            "example": "LEVEL_2",
            "description": "Survey type (LEVEL_2, LEVEL_3, VALUATION, SNAGGING, REINSPECTION, OTHER)",
            "default": "LEVEL_2"
          },
          "jobRef": {
            "type": "string",
            "example": "JOB-2024-001234",
            "description": "Job reference number"
          },
          "clientName": {
            "type": "string",
            "example": "John Smith",
            "description": "Client name for the survey"
          },
          "parentSurveyId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "description": "Parent survey ID for reinspections"
          },
          "sections": {
            "description": "Sections within the survey",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateSectionDto"
            }
          }
        },
        "required": [
          "title",
          "propertyAddress"
        ]
      },
...
      "SectionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "surveyId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "title": {
            "type": "string",
            "example": "Roof Inspection"
          },
          "order": {
            "type": "number",
            "example": 1
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "surveyId",
          "title",
          "order",
          "createdAt",
          "updatedAt"
        ]
      },
      "SurveyResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "title": {
            "type": "string",
            "example": "Level 2 Home Survey - 123 Main Street"
          },
          "propertyAddress": {
            "type": "string",
            "example": "123 Main Street, London, SW1A 1AA"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "IN_PROGRESS",
              "PAUSED",
              "COMPLETED",
              "PENDING_REVIEW",
              "APPROVED",
              "REJECTED"
            ],
            "example": "DRAFT"
          },
          "type": {
            "type": "string",
            "enum": [
              "LEVEL_2",
              "LEVEL_3",
              "VALUATION",
              "SNAGGING",
              "REINSPECTION",
              "OTHER"
            ],
            "example": "LEVEL_2"
          },
          "jobRef": {
            "type": "string",
            "example": "JOB-2024-001234"
          },
          "clientName": {
            "type": "string",
            "example": "John Smith"
          },
          "parentSurveyId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "userId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SectionResponseDto"
            }
          }
        },
        "required": [
          "id",
          "title",
          "propertyAddress",
          "status",
          "userId",
          "createdAt",
          "updatedAt",
          "sections"
        ]
      },
...
      "UpdateSectionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Roof Inspection - Updated",
            "description": "Title of the section"
          },
          "order": {
            "type": "number",
            "example": 2,
            "description": "Order of the section within the survey"
          }
        }
      },
      "UpdateSurveyDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Level 2 Home Survey - 123 Main Street",
            "description": "Title of the survey"
          },
          "propertyAddress": {
            "type": "string",
            "example": "123 Main Street, London, SW1A 1AA",
            "description": "Property address being surveyed"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "IN_PROGRESS",
              "PAUSED",
              "COMPLETED",
              "PENDING_REVIEW",
              "APPROVED",
              "REJECTED"
            ],
            "example": "COMPLETED",
            "description": "Survey status"
          },
          "type": {
            "type": "string",
            "enum": [
              "LEVEL_2",
              "LEVEL_3",
              "VALUATION",
              "SNAGGING",
              "REINSPECTION",
              "OTHER"
            ],
            "example": "LEVEL_2",
            "description": "Survey type"
          },
          "jobRef": {
            "type": "string",
            "example": "JOB-2024-001234",
            "description": "Job reference number"
          },
          "clientName": {
            "type": "string",
            "example": "John Smith",
            "description": "Client name for the survey"
          },
          "parentSurveyId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "description": "Parent survey ID for reinspections"
          },
          "sections": {
            "description": "Sections within the survey (replaces existing sections)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateSectionDto"
            }
          }
        }
      },
      "DeleteSurveyResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "success",
          "id",
          "deletedAt"
        ]
      },
...
    }
  }
}